refactor: changed structure to packages feature: decoration class added docs: javadoc comments added
