--- src/main/java/Application.java
@@ -1,5 +1,0 @@
-class Application {
-    public static void main(String[] args) {
-        ConsolePrinter.print("Hello world!");
-    }
-}
