+++ src/main/java/com/tcs/edu/Application.java
@@ -1,0 +1,11 @@
+package com.tcs.edu;
+
+import com.tcs.edu.decorator.TimestampMessageDecorator;
+import com.tcs.edu.printer.ConsolePrinter;
+
+class Application {
+    public static void main(String[] args) {
+
+        ConsolePrinter.print(TimestampMessageDecorator.decorate("Hello world!"));
+    }
+}
