+++ src/main/java/com/tcs/edu/printer/ConsolePrinter.java
@@ -1,0 +1,17 @@
+package com.tcs.edu.printer;
+
+/**
+ * Class for printing a message
+ * @author Alexey Knyazev
+ */
+
+public class ConsolePrinter {
+    /**
+     * Prints given string
+     * @param message string to print
+     */
+    public static void print(String message) {
+
+        System.out.println(message);
+    }
+}
